ep11-more-routing/bower_components/web-animations-next/test/blink/add-keyframes.html

<!--
 Copyright 2017 Google Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
      http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-->

<!doctype html>
<title>Web Animations API: Add Keyframe tests</title>
<script src="testharness/testharness.js"></script>
<script src="testharness/testharnessreport.js"></script>
<div id='log'></div>
<div id='div'></div>
<script>

test(function() {
  assert_throws({name: 'NotSupportedError'}, function() {
    div.animate([{height: '100px', composite: 'add'}, {height: '200px', composite: 'add'}], 1);
  }, 'Add animations should cause an exception.');
},
'Add animation tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#the-unaccumulated-animation-value-of-a-keyframe-animation-effect',
  assert: 'Only keyframes without add compositing considered',
  author: 'Shane Stephens'
});

test(function() {
  assert_throws({name: 'NotSupportedError'}, function() {
    div.animate([{height: '100px'}, {height: '200px', composite: 'add'}], 1);
  }, 'Animations with add keyframes should cause an exception.');
},
'Hybrid animation tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#the-unaccumulated-animation-value-of-a-keyframe-animation-effect',
  assert: 'Only keyframes without add compositing are considered',
  author: 'Shane Stephens'
});

</script>
